Very Simple Player Health Bar
I was going to delete this script, but I want to keep it because it works for the time being. It feeds into another script for player health. You will see this one basically just defines the players health bar. Not much too it. Will probably refactor in the future, but until the changes can be made. It will help us with attack features.