Wizard model Import & Refactoring
Adding new anims for Wizard, as well as an anim controller, scene updates, folder  organization, changes in Player Class Data script for Wizard, deleting old prototype animations.

-- SCENE UPDATES ARE NOT DONE.
If you go to GavinsScene there will be errors. I needed to push these changes for the models to be in github though.